requirements.txt
apache-airflow-providers-common-sql==1.14.1
apache-airflow-providers-http==4.12.0
apache-airflow-providers-amazon==8.25.0 
apache-airflow-client==2.9.0
apache-airflow-providers-postgres==5.11.2
apache-airflow==2.9.2